move basic integration test to separate folder
