+++ framework/integration/init.go
@@ -17,10 +17,6 @@ import (
 	"github.com/giantswarm/operatorkit/informer"
 )
 
-const (
-	namespace = "finalizer-integration-test"
-)
-
 var (
 	err error
 
@@ -48,7 +44,7 @@ func newK8sClient() (kubernetes.Interface, error) {
 	return k8sClient, nil
 }
 
-func newFramework(name string) (*framework.Framework, error) {
+func NewFramework(name, namespace string) (*framework.Framework, error) {
 	logger, err := micrologger.New(micrologger.Config{})
 	if err != nil {
 		return nil, microerror.Mask(err)
@@ -107,8 +103,8 @@ func newFramework(name string) (*framework.Framework, error) {
 	return f, nil
 }
 
-func mustSetup() {
-	mustTeardown()
+func MustSetup(namespace string) {
+	MustTeardown(namespace)
 
 	ns := &corev1.Namespace{
 		TypeMeta: metav1.TypeMeta{
@@ -127,7 +123,7 @@ func mustSetup() {
 	}
 }
 
-func mustTeardown() {
+func MustTeardown(namespace string) {
 	err := k8sClient.CoreV1().Namespaces().Delete(namespace, nil)
 	if errors.IsNotFound(err) {
 		// fall though
